+++ src/test/java/stepDefinition/BaseStep.java
@@ -1,0 +1,33 @@
+package stepDefinition;
+
+import io.cucumber.java.After;
+import io.cucumber.java.Before;
+import io.github.bonigarcia.wdm.WebDriverManager;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.interactions.Actions;
+import org.openqa.selenium.support.ui.WebDriverWait;
+
+import java.time.Duration;
+
+public class BaseStep {
+
+    protected static WebDriver driver = null;
+
+    protected WebDriverWait wait = null;
+    protected Actions actions = null;
+
+    @Before
+    public void openBrowser(){
+        WebDriverManager.chromedriver().setup();
+        driver = new ChromeDriver();
+        wait = new WebDriverWait(driver, Duration.ofSeconds(120));
+        actions = new Actions(driver);
+    }
+
+    @After
+    public void closeBrowser() {
+        driver.quit();
+    }
+
+}
