+++ src/test/java/stepDefinition/CommonSteps.java
@@ -1,0 +1,44 @@
+package stepDefinition;
+
+import io.cucumber.java.After;
+import io.cucumber.java.Before;
+import io.cucumber.java.en.And;
+import io.cucumber.java.en.When;
+import io.github.bonigarcia.wdm.WebDriverManager;
+import io.github.bonigarcia.wdm.managers.ChromeDriverManager;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.interactions.Actions;
+import org.openqa.selenium.support.ui.WebDriverWait;
+import org.testng.annotations.*;
+import pages.LoginPage;
+
+import java.time.Duration;
+
+public class CommonSteps  {
+//    static WebDriver driver = BaseStep.driver;
+
+    @And("I open Login Page")
+    public void iOpenLoginPage() {
+        BaseStep.driver.get("http://bbb.testpro.io");
+    }
+    @When("I enter email {string}")
+    public void iEnterEmail(String email) {
+        LoginPage loginPage = new LoginPage(BaseStep.driver);
+        loginPage.provideEmail(email);
+    }
+
+    @And("I enter password {string}")
+    public void iEnterPassword(String password) {
+        LoginPage loginPage = new LoginPage(BaseStep.driver);
+        loginPage.providePassword(password);
+    }
+
+    @And("I submit")
+    public void iSubmit() {
+        LoginPage loginPage = new LoginPage(BaseStep.driver);
+        loginPage.clickSubmitBtn();
+    }
+
+
+}
