+++ src/test/java/stepDefinition/LoginStepDefinition.java
@@ -19,54 +19,19 @@ import pages.LoginPage;
 
 import java.time.Duration;
 
-public class LoginStepDefinition {
-    protected WebDriver driver;
-    protected WebDriverWait wait;
+public class LoginStepDefinition  {
 
-    @Before
-    public void openBrowser(){
-        ChromeDriverManager.getInstance().setup();
-        WebDriverManager.chromedriver().setup();
-        driver = new ChromeDriver();
-        wait = new WebDriverWait(driver, Duration.ofSeconds(120));
 
-    }
-    @After
-    public void closeBrowser() {
 
-        driver.quit();
-    }
 
-    @And("I open Login Page")
-    public void iOpenLoginPage() {
-        driver.get("http://bbb.testpro.io");
 
-    }
 
-    @When("I enter email {string}")
-    public void iEnterEmail(String email) {
-        LoginPage loginPage = new LoginPage(driver);
-        loginPage.provideEmail(email);
-    }
 
-    @And("I enter password {string}")
-    public void iEnterPassword(String password) {
-        LoginPage loginPage = new LoginPage(driver);
-        loginPage.providePassword(password);
-     }
-
-    @And("I submit")
-    public void iSubmit() {
-        LoginPage loginPage = new LoginPage(driver);
-        loginPage.clickSubmitBtn();
-      }
-
-    @Then("I am logged in")
-    public void iAmLoggedIn() {
-                Assert.assertTrue(
-                wait.until(ExpectedConditions
-                        .elementToBeClickable(By.cssSelector("img.avatar"))).isDisplayed());
-    }
+//    @Then("I am logged in")
+//    public void iAmLoggedIn() {
+//                Assert.assertTrue(wait.until(ExpectedConditions
+//                        .elementToBeClickable(By.cssSelector("img.avatar"))).isDisplayed());
+//    }
 
 //    @When("I enter email <username>")
   //  public void iEnterEmailUsername(String username) {
@@ -80,12 +45,13 @@ public class LoginStepDefinition {
  //       loginPage.providePassword(password);
 //    }
 
-    @Then("I get an error message")
-    public void iGetAnErrorMessage() {
-              Assert.assertTrue(wait.until(ExpectedConditions.elementToBeClickable(By.cssSelector(".error"))).isDisplayed());
-    }
-    @Then("I get Enter Email and Password")
-    public void messageEnterEmailAndPassword(){
-        Assert.assertTrue(wait.until(ExpectedConditions.elementToBeClickable(By.cssSelector("[type='email']"))).isDisplayed());
-    }
+//    @Then("I get an error message")
+//    public void iGetAnErrorMessage() {
+//              Assert.assertTrue(wait.until(ExpectedConditions.elementToBeClickable(By.cssSelector(".error"))).isDisplayed());
+//            //lets put this on POM/page objects
+//    }
+//    @Then("I get Enter Email and Password")
+//    public void messageEnterEmailAndPassword(){
+//        Assert.assertTrue(wait.until(ExpectedConditions.elementToBeClickable(By.cssSelector("[type='email']"))).isDisplayed());
+//    }
 }
